requirements.txt
markdown
pdfkit
python-dotenv
crewai
crewai-tools
langchain-community
